@@ -1,159 +1,107 @@
 package com.baeldung.spring.data.cassandra.repository;
 
-import com.baeldung.spring.data.cassandra.config.CassandraConfig;
-import com.baeldung.spring.data.cassandra.model.Book;
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.Session;
-import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.datastax.driver.core.querybuilder.Select;
-import com.datastax.driver.core.utils.UUIDs;
-import com.google.common.collect.ImmutableSet;
-import org.apache.cassandra.exceptions.ConfigurationException;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.apache.thrift.transport.TTransportException;
-import org.cassandraunit.utils.EmbeddedCassandraServerHelper;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cassandra.core.cql.CqlIdentifier;
-import org.springframework.data.cassandra.core.CassandraAdminOperations;
-import org.springframework.data.cassandra.core.CassandraOperations;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-
-import static junit.framework.TestCase.assertNull;
+import static com.datastax.oss.driver.api.querybuilder.QueryBuilder.literal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-/**
- * Live test for Cassandra testing.
- *
- * This can be converted to IntegrationTest once cassandra-unit tests can be executed in parallel and
- * multiple test servers started as part of test suite.
- *
- * Open cassandra-unit issue for parallel execution: https://github.com/jsevellec/cassandra-unit/issues/155
- */
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = CassandraConfig.class)
-public class CassandraTemplateLiveTest {
-    private static final Log LOGGER = LogFactory.getLog(CassandraTemplateLiveTest.class);
+import java.util.Collections;
+import java.util.List;
+import java.util.Set;
 
-    public static final String KEYSPACE_CREATION_QUERY = "CREATE KEYSPACE IF NOT EXISTS testKeySpace " + "WITH replication = { 'class': 'SimpleStrategy', 'replication_factor': '3' };";
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.data.cassandra.core.CassandraAdminOperations;
+import org.testcontainers.containers.CassandraContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
-    public static final String KEYSPACE_ACTIVATE_QUERY = "USE testKeySpace;";
+import com.baeldung.spring.data.cassandra.model.Book;
+import com.datastax.oss.driver.api.core.CqlIdentifier;
+import com.datastax.oss.driver.api.core.CqlSession;
+import com.datastax.oss.driver.api.core.cql.SimpleStatement;
+import com.datastax.oss.driver.api.core.uuid.Uuids;
+import com.datastax.oss.driver.api.querybuilder.QueryBuilder;
 
-    public static final String DATA_TABLE_NAME = "book";
+@Testcontainers
+@SpringBootTest(classes = CassandraTestConfiguration.class)
+class CassandraTemplateLiveTest {
 
-    @Autowired
-    private CassandraAdminOperations adminTemplate;
+    private static final String KEYSPACE_NAME = "testKeySpace";
+    private static final String KEYSPACE_CREATION_QUERY = "CREATE KEYSPACE IF NOT EXISTS testKeySpace " + "WITH replication = { 'class': 'SimpleStrategy', 'replication_factor': '1' };";
+    private static final String KEYSPACE_ACTIVATE_QUERY = "USE " + KEYSPACE_NAME + ";";
 
-    @Autowired
-    private CassandraOperations cassandraTemplate;
+    private static final String DATA_TABLE_NAME = "book";
 
-    @BeforeClass
-    public static void startCassandraEmbedded() throws InterruptedException, TTransportException, ConfigurationException, IOException {
-        EmbeddedCassandraServerHelper.startEmbeddedCassandra();
-        final Cluster cluster = Cluster.builder().addContactPoints("127.0.0.1").withPort(9142).build();
-        LOGGER.info("Server Started at 127.0.0.1:9142... ");
-        final Session session = cluster.connect();
-        session.execute(KEYSPACE_CREATION_QUERY);
-        session.execute(KEYSPACE_ACTIVATE_QUERY);
-        LOGGER.info("KeySpace created and activated.");
-        Thread.sleep(5000);
+    @Container
+    private static final CassandraContainer<?> cassandraContainer = new CassandraContainer<>("cassandra:4.1.8").withExposedPorts(9042);
+
+    @BeforeAll
+    static void setupCassandraConnectionProperties() {
+        System.setProperty("spring.cassandra.keyspace-name", KEYSPACE_NAME);
+        System.setProperty("spring.cassandra.contact-points", cassandraContainer.getContainerIpAddress());
+        System.setProperty("spring.cassandra.port", String.valueOf(cassandraContainer.getMappedPort(9042)));
+
+        try (CqlSession session = CqlSession.builder()
+            .addContactPoint(cassandraContainer.getContactPoint())
+            .withLocalDatacenter(cassandraContainer.getLocalDatacenter())
+            .build()) {
+
+            session.execute(KEYSPACE_CREATION_QUERY);
+            session.execute(KEYSPACE_ACTIVATE_QUERY);
+        }
     }
 
-    @Before
-    public void createTable() {
-        adminTemplate.createTable(true, CqlIdentifier.cqlId(DATA_TABLE_NAME), Book.class, new HashMap<>());
+    @Autowired
+    private CassandraAdminOperations cassandraTemplate;
+
+    @BeforeEach
+    void createTable() {
+        cassandraTemplate.createTable(true, CqlIdentifier.fromCql(DATA_TABLE_NAME), Book.class, Collections.emptyMap());
     }
 
     @Test
-    public void whenSavingBook_thenAvailableOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
+    void whenSavingBook_thenAvailableOnRetrieval() {
+        Book javaBook = new Book(Uuids.timeBased(), "Head First Java", "O'Reilly Media", Collections.singleton("Computer"));
+
         cassandraTemplate.insert(javaBook);
-        final Select select = QueryBuilder.select().from("book").where(QueryBuilder.eq("title", "Head First Java")).and(QueryBuilder.eq("publisher", "O'Reilly Media")).limit(10);
-        final Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
+
+        SimpleStatement select = QueryBuilder.selectFrom(DATA_TABLE_NAME)
+            .all()
+            .whereColumn("title")
+            .isEqualTo(literal("Head First Java"))
+            .allowFiltering()
+            .build();
+
+        Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
         assertEquals(javaBook.getId(), retrievedBook.getId());
     }
 
     @Test
-    public void whenSavingBooks_thenAllAvailableOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        final Book dPatternBook = new Book(UUIDs.timeBased(), "Head Design Patterns", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        final List<Book> bookList = new ArrayList<>();
-        bookList.add(javaBook);
-        bookList.add(dPatternBook);
-        cassandraTemplate.insert(bookList);
+    void whenSavingBooks_thenAllAvailableOnRetrieval() {
+        List<Book> books = List.of(new Book(Uuids.timeBased(), "Head First Java", "O'Reilly Media", Set.of("Computer")), new Book(Uuids.timeBased(), "Clean Code", "Pearson", Set.of("Software")));
 
-        final Select select = QueryBuilder.select().from("book").limit(10);
-        final List<Book> retrievedBooks = cassandraTemplate.select(select, Book.class);
-        assertThat(retrievedBooks.size(), is(2));
-        assertEquals(javaBook.getId(), retrievedBooks.get(0).getId());
-        assertEquals(dPatternBook.getId(), retrievedBooks.get(1).getId());
+        for (Book book : books) {
+            cassandraTemplate.insert(book);
+        }
+
+        SimpleStatement select = QueryBuilder.selectFrom(DATA_TABLE_NAME)
+            .all()
+            .build();
+
+        List<Book> retrieved = cassandraTemplate.select(select, Book.class);
+        assertThat(retrieved.size(), is(3));
     }
 
-    @Test
-    public void whenUpdatingBook_thenShouldUpdatedOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        cassandraTemplate.insert(javaBook);
-        final Select select = QueryBuilder.select().from("book").limit(10);
-        final Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
-        retrievedBook.setTags(ImmutableSet.of("Java", "Programming"));
-        cassandraTemplate.update(retrievedBook);
-        final Book retrievedUpdatedBook = cassandraTemplate.selectOne(select, Book.class);
-        assertEquals(retrievedBook.getTags(), retrievedUpdatedBook.getTags());
-    }
 
-    @Test
-    public void whenDeletingASelectedBook_thenNotAvailableOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "OReilly Media", ImmutableSet.of("Computer", "Software"));
-        cassandraTemplate.insert(javaBook);
-        cassandraTemplate.delete(javaBook);
-        final Select select = QueryBuilder.select().from("book").limit(10);
-        final Book retrievedUpdatedBook = cassandraTemplate.selectOne(select, Book.class);
-        assertNull(retrievedUpdatedBook);
-    }
-
-    @Test
-    public void whenDeletingAllBooks_thenNotAvailableOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        final Book dPatternBook = new Book(UUIDs.timeBased(), "Head Design Patterns", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        cassandraTemplate.insert(javaBook);
-        cassandraTemplate.insert(dPatternBook);
-        cassandraTemplate.deleteAll(Book.class);
-        final Select select = QueryBuilder.select().from("book").limit(10);
-        final Book retrievedUpdatedBook = cassandraTemplate.selectOne(select, Book.class);
-        assertNull(retrievedUpdatedBook);
-    }
-
-    @Test
-    public void whenAddingBooks_thenCountShouldBeCorrectOnRetrieval() {
-        final Book javaBook = new Book(UUIDs.timeBased(), "Head First Java", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        final Book dPatternBook = new Book(UUIDs.timeBased(), "Head Design Patterns", "O'Reilly Media", ImmutableSet.of("Computer", "Software"));
-        cassandraTemplate.insert(javaBook);
-        cassandraTemplate.insert(dPatternBook);
-        final long bookCount = cassandraTemplate.count(Book.class);
-        assertEquals(2, bookCount);
-    }
-
-    @After
-    public void dropTable() {
-        adminTemplate.dropTable(CqlIdentifier.cqlId(DATA_TABLE_NAME));
-    }
-
-    @AfterClass
-    public static void stopCassandraEmbedded() {
-        EmbeddedCassandraServerHelper.cleanEmbeddedCassandra();
+    @AfterAll
+    static void tearDown() {
+        if (cassandraContainer != null) {
+            cassandraContainer.stop();
+        }
     }
 }
