@@ -1,126 +1,132 @@
 package com.baeldung.spring.data.cassandra.repository;
 
-import com.baeldung.spring.data.cassandra.config.CassandraConfig;
-import com.baeldung.spring.data.cassandra.model.Book;
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.Session;
-import com.datastax.driver.core.querybuilder.Insert;
-import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.datastax.driver.core.querybuilder.Select;
-import com.datastax.driver.core.utils.UUIDs;
-import com.google.common.collect.ImmutableSet;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.cassandraunit.utils.EmbeddedCassandraServerHelper;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cassandra.core.cql.CqlIdentifier;
-import org.springframework.data.cassandra.core.CassandraAdminOperations;
-import org.springframework.data.cassandra.core.CassandraOperations;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
-import static junit.framework.TestCase.assertEquals;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.data.cassandra.core.CassandraAdminOperations;
+import org.testcontainers.containers.CassandraContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+import com.baeldung.spring.data.cassandra.model.Book;
+import com.datastax.driver.core.utils.UUIDs;
+import com.datastax.oss.driver.api.core.CqlIdentifier;
+import com.datastax.oss.driver.api.core.CqlSession;
+import com.datastax.oss.driver.api.core.cql.SimpleStatement;
+import com.datastax.oss.driver.api.querybuilder.QueryBuilder;
+import com.datastax.oss.driver.api.querybuilder.insert.RegularInsert;
+import com.google.common.collect.ImmutableSet;
 
 /**
  * Live test for Cassandra testing.
- *
- * This can be converted to IntegrationTest once cassandra-unit tests can be executed in parallel and
- * multiple test servers started as part of test suite.
- *
- * Open cassandra-unit issue for parallel execution: https://github.com/jsevellec/cassandra-unit/issues/155
  */
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = CassandraConfig.class)
-public class CqlQueriesLiveTest {
-    private static final Log LOGGER = LogFactory.getLog(CqlQueriesLiveTest.class);
+@Testcontainers
+@SpringBootTest(classes = CassandraTestConfiguration.class)
+class CqlQueriesLiveTest {
 
-    public static final String KEYSPACE_CREATION_QUERY = "CREATE KEYSPACE IF NOT EXISTS testKeySpace " + "WITH replication = { 'class': 'SimpleStrategy', 'replication_factor': '3' };";
+    private static final String KEYSPACE_NAME = "testKeySpace";
 
-    public static final String KEYSPACE_ACTIVATE_QUERY = "USE testKeySpace;";
-
-    public static final String DATA_TABLE_NAME = "book";
-
-    @Autowired
-    private CassandraAdminOperations adminTemplate;
+    private static final String KEYSPACE_CREATION_QUERY = "CREATE KEYSPACE IF NOT EXISTS testKeySpace " + "WITH replication = { 'class': 'SimpleStrategy', 'replication_factor': '1' };";
+    private static final String KEYSPACE_ACTIVATE_QUERY = "USE testKeySpace;";
+    private static final String DATA_TABLE_NAME = "book";
 
     @Autowired
-    private CassandraOperations cassandraTemplate;
+    private CassandraAdminOperations cassandraTemplate;
 
-    @BeforeClass
-    public static void startCassandraEmbedded() throws Exception {
-        EmbeddedCassandraServerHelper.startEmbeddedCassandra(25000);
-        final Cluster cluster = Cluster.builder().addContactPoints("127.0.0.1").withPort(9142).build();
-        LOGGER.info("Server Started at 127.0.0.1:9142... ");
-        final Session session = cluster.connect();
-        session.execute(KEYSPACE_CREATION_QUERY);
-        session.execute(KEYSPACE_ACTIVATE_QUERY);
-        LOGGER.info("KeySpace created and activated.");
-        Thread.sleep(5000);
+    @Container
+    private static final CassandraContainer<?> cassandraContainer = new CassandraContainer<>("cassandra:4.1.8").withExposedPorts(9042);
+
+    @BeforeAll
+    static void setupCassandraConnectionProperties() {
+        System.setProperty("spring.cassandra.keyspace-name", KEYSPACE_NAME);
+        System.setProperty("spring.cassandra.contact-points", cassandraContainer.getContainerIpAddress());
+        System.setProperty("spring.cassandra.port", String.valueOf(cassandraContainer.getMappedPort(9042)));
+
+        try (CqlSession session = CqlSession.builder()
+            .addContactPoint(cassandraContainer.getContactPoint())
+            .withLocalDatacenter(cassandraContainer.getLocalDatacenter())
+            .build()) {
+
+            session.execute(KEYSPACE_CREATION_QUERY);
+            session.execute(KEYSPACE_ACTIVATE_QUERY);
+        }
     }
 
-    @Before
-    public void createTable() {
-        adminTemplate.createTable(true, CqlIdentifier.cqlId(DATA_TABLE_NAME), Book.class, new HashMap<>());
+    @AfterAll
+    static void tearDown() {
+        if (cassandraContainer != null) {
+            cassandraContainer.stop();
+        }
+    }
+
+    @BeforeEach
+    void createTable() {
+        cassandraTemplate.createTable(true, CqlIdentifier.fromCql(DATA_TABLE_NAME), Book.class, Collections.emptyMap());
     }
 
     @Test
-    public void whenSavingBook_thenAvailableOnRetrieval_usingQueryBuilder() {
+    void whenSavingBook_thenAvailableOnRetrieval_usingQueryBuilder() {
         final UUID uuid = UUIDs.timeBased();
-        final Insert insert = QueryBuilder.insertInto(DATA_TABLE_NAME).value("id", uuid).value("title", "Head First Java").value("publisher", "OReilly Media").value("tags", ImmutableSet.of("Software"));
-        cassandraTemplate.execute(insert);
-        final Select select = QueryBuilder.select().from("book").limit(10);
-        final Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
+
+        final RegularInsert insert = QueryBuilder.insertInto(DATA_TABLE_NAME)
+            .value("id", QueryBuilder.literal(uuid))
+            .value("title", QueryBuilder.literal("Head First Java"))
+            .value("publisher", QueryBuilder.literal("OReilly Media"))
+            .value("tags", QueryBuilder.literal(ImmutableSet.of("Software")));
+
+        cassandraTemplate.getCqlOperations()
+            .execute(insert.asCql());
+
+        final SimpleStatement select = QueryBuilder.selectFrom(DATA_TABLE_NAME)
+            .all()
+            .limit(10)
+            .build();
+        Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
+
         assertEquals(uuid, retrievedBook.getId());
     }
 
     @Test
-    public void whenSavingBook_thenAvailableOnRetrieval_usingCQLStatements() {
+    void whenSavingBook_thenAvailableOnRetrieval_usingCQLStatements() {
         final UUID uuid = UUIDs.timeBased();
         final String insertCql = "insert into book (id, title, publisher, tags) values " + "(" + uuid + ", 'Head First Java', 'OReilly Media', {'Software'})";
-        cassandraTemplate.execute(insertCql);
-        final Select select = QueryBuilder.select().from("book").limit(10);
+        cassandraTemplate.getCqlOperations()
+            .execute(insertCql);
+        SimpleStatement select = QueryBuilder.selectFrom(DATA_TABLE_NAME)
+            .all()
+            .limit(10)
+            .build();
         final Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
         assertEquals(uuid, retrievedBook.getId());
     }
 
     @Test
-    public void whenSavingBook_thenAvailableOnRetrieval_usingPreparedStatements() throws InterruptedException {
+    void whenSavingBook_thenAvailableOnRetrieval_usingPreparedStatements() {
         final UUID uuid = UUIDs.timeBased();
         final String insertPreparedCql = "insert into book (id, title, publisher, tags) values (?, ?, ?, ?)";
-        final List<Object> singleBookArgsList = new ArrayList<>();
-        final List<List<?>> bookList = new ArrayList<>();
-        singleBookArgsList.add(uuid);
-        singleBookArgsList.add("Head First Java");
-        singleBookArgsList.add("OReilly Media");
-        singleBookArgsList.add(ImmutableSet.of("Software"));
-        bookList.add(singleBookArgsList);
-        cassandraTemplate.ingest(insertPreparedCql, bookList);
-        // This may not be required, just added to avoid any transient issues
-        Thread.sleep(5000);
-        final Select select = QueryBuilder.select().from("book");
+
+        List<List<Object>> bookArgsList = new ArrayList<>();
+        bookArgsList.add(List.of(uuid, "Head First Java", "OReilly Media", ImmutableSet.of("Software")));
+
+        for (List<Object> list : bookArgsList) {
+            cassandraTemplate.getCqlOperations()
+                .execute(insertPreparedCql, list.toArray());
+        }
+
+        final SimpleStatement select = QueryBuilder.selectFrom(DATA_TABLE_NAME)
+            .all()
+            .build();
         final Book retrievedBook = cassandraTemplate.selectOne(select, Book.class);
         assertEquals(uuid, retrievedBook.getId());
     }
-
-    @After
-    public void dropTable() {
-        adminTemplate.dropTable(CqlIdentifier.cqlId(DATA_TABLE_NAME));
-    }
-
-    @AfterClass
-    public static void stopCassandraEmbedded() {
-        EmbeddedCassandraServerHelper.cleanEmbeddedCassandra();
-    }
-
 }
