@@ -1,14 +1,15 @@
 package com.baeldung.spring.data.cassandra.repository;
 
-import com.baeldung.spring.data.cassandra.model.Book;
+import java.util.UUID;
+
 import org.springframework.data.cassandra.repository.CassandraRepository;
-import org.springframework.data.cassandra.repository.Query;
 import org.springframework.stereotype.Repository;
 
-@Repository
-public interface BookRepository extends CassandraRepository<Book> {
+import com.baeldung.spring.data.cassandra.model.Book;
 
-    @Query("select * from book where title = ?0 and publisher=?1")
+@Repository
+public interface BookRepository extends CassandraRepository<Book, UUID> {
+
     Iterable<Book> findByTitleAndPublisher(String title, String publisher);
 
 }
